requirements.txt
python-dotenv>=1.0.1
requests>=2.32.3
psycopg2-binary==2.9.9
pandas>=2.2
numpy>=1.26
psycopg2-binary>=2.9
python-dotenv>=1.0
